move serialized_component_batch to separate file
