+++ rerun_cpp/src/rerun/serialized_component_batch.hpp
@@ -1,0 +1,17 @@
+#include "data_cell.hpp"
+
+namespace rerun {
+    /// An arrow serialized `ComponentBatch`.
+    struct SerializedComponentBatch {
+        SerializedComponentBatch() = default;
+
+        SerializedComponentBatch(size_t _num_instances, DataCell _data_cell)
+            : num_instances(_num_instances), data_cell(std::move(_data_cell)) {}
+
+        /// How many components were serialized.
+        size_t num_instances;
+
+        /// The underlying data.
+        DataCell data_cell;
+    };
+} // namespace rerun
